Enable -Xlint and -Werror
Linting on 2.13 such as required empty args,
inferred Any in sameElements of Array[_],
Timeout is double-valued, avoid Any.equals,
some elements unused, deprecations.
Allow JavaConverters.

Can't use nowarn and unused annotations
under 2.11/2.12.

+++ munit/js/src/main/scala/munit/internal/PlatformCompat.scala
@@ -30,7 +30,7 @@ object PlatformCompat {
       ec: ExecutionContext
   ): Future[T] = {
     val onComplete = Promise[T]()
-    val timeoutHandle = timers.setTimeout(duration.toMillis) {
+    val timeoutHandle = timers.setTimeout(duration.toMillis.toDouble) {
       onComplete.tryFailure(
         new TimeoutException(s"test timed out after $duration")
       )
